Fix wand monkey bug which prevented magic defense resists

+++ Source/ACE.Server/WorldObjects/WorldObject_Magic.cs
@@ -170,7 +170,7 @@ namespace ACE.Server.WorldObjects
 
                     var schoolSkill = wielder.GetCreatureSkill(spell.School).Current;
 
-                    var arcaneLore = wielder.GetCreatureSkill(Skill.ArcaneLore).Current;
+                    var arcaneLore = (int)wielder.GetCreatureSkill(Skill.ArcaneLore).Current;
                     var arcaneLoreMod = Math.Max(1.0f + Math.Max(arcaneLore - 55, 0) * 0.002f, 1.0f);
                     var ArcaneLoreModifiedSpellcraft = (uint)(magicSkill * arcaneLoreMod);
